+++ compiler/libsolidity/ast/Types.cpp
@@ -3466,8 +3466,11 @@ std::string FunctionType::richIdentifier() const
   case Kind::GoshCNVRTSHELLQ: id += "goshcnvrtshellq"; break;
   case Kind::GoshMINTSHELL: id += "goshmintshell"; break;
   case Kind::GoshCALCBKREWARD: id += "goshcalcbkreward"; break;
+  case Kind::GoshCALCBMREWARD: id += "goshcalcbmreward"; break;
   case Kind::GoshCALCMINSTAKE: id += "goshcalcminstake"; break;
+  case Kind::GoshCALCMINSTAKEBM: id += "goshcalcminstakebm"; break;
   case Kind::GoshCALCBKREWARDADJ: id += "goshcalcbkrewardadj"; break;
+  case Kind::GoshCALCBMREWARDADJ: id += "goshcalcbmrewardadj"; break;
   case Kind::GoshCALCREPCOEF: id += "goshcalcrepcoef"; break;
 	}
 	id += "_" + stateMutabilityToString(m_stateMutability);
@@ -5707,6 +5710,18 @@ MemberList::MemberMap MagicType::nativeMembers(ASTNode const*) const
 		)});
 
 		members.push_back({
+			"calcbmreward",
+			TypeProvider::function(
+				{TypeProvider::uint128(), TypeProvider::uint128(), TypeProvider::uint128(), TypeProvider::uint128()}, 
+				{TypeProvider::uint128()},
+				{{}, {}, {}, {}}, 
+				{{}},
+				FunctionType::Kind::GoshCALCBMREWARD,
+				StateMutability::Pure,
+				nullptr, FunctionType::Options::withArbitraryParameters()
+		)});
+
+		members.push_back({
 			"calcbkrewardadj",
 			TypeProvider::function(
 				{TypeProvider::uint128(), TypeProvider::uint128(), TypeProvider::uint128(), TypeProvider::uint128(), TypeProvider::uint128()}, 
@@ -5714,6 +5729,18 @@ MemberList::MemberMap MagicType::nativeMembers(ASTNode const*) const
 				{{}, {}, {}, {}, {}}, 
 				{{}},
 				FunctionType::Kind::GoshCALCBKREWARDADJ,
+				StateMutability::Pure,
+				nullptr, FunctionType::Options::withArbitraryParameters()
+		)});
+
+		members.push_back({
+			"calcbmrewardadj",
+			TypeProvider::function(
+				{TypeProvider::uint128(), TypeProvider::uint128(), TypeProvider::uint128(), TypeProvider::uint128()}, 
+				{TypeProvider::uint128()},
+				{{}, {}, {}, {}}, 
+				{{}},
+				FunctionType::Kind::GoshCALCBMREWARDADJ,
 				StateMutability::Pure,
 				nullptr, FunctionType::Options::withArbitraryParameters()
 		)});
@@ -5738,6 +5765,18 @@ MemberList::MemberMap MagicType::nativeMembers(ASTNode const*) const
 				{{}, {}, {}, {}}, 
 				{{}},
 				FunctionType::Kind::GoshCALCMINSTAKE,
+				StateMutability::Pure,
+				nullptr, FunctionType::Options::withArbitraryParameters()
+		)});
+
+		members.push_back({
+			"calcminstakebm",
+			TypeProvider::function(
+				{TypeProvider::uint128(), TypeProvider::uint128()}, 
+				{TypeProvider::uint128()},
+				{{}, {}}, 
+				{{}},
+				FunctionType::Kind::GoshCALCMINSTAKEBM,
 				StateMutability::Pure,
 				nullptr, FunctionType::Options::withArbitraryParameters()
 		)});