+++ Changelog.md
@@ -1,10 +1,15 @@
+### 0.77.1 (2025-05-21)
+
+Compiler features:
+ * Supported `CALCBMREWARD`, `CALCBMREWARDADJ`, `CALCMINSTAKEBM`instruction.
+
 ### 0.77.0 (2025-02-19)
 
 Compiler features:
  * Deprecated +2 and 64 flag in transfer
  * Left only 0 and +1 flag in RAWRESERVE 
  * Supported `CALCBKREWARDADJ`instruction.
- * Updated `CACMINSTAKE`, `CALCBKREWARD` instructions.
+ * Updated `CALCMINSTAKE`, `CALCBKREWARD` instructions.
 
 ### 0.76.1 (2024-12-09)